Move utility functions to HddsUtils
